fix: compilation errors for AGenUI Playground build
- Add missing import android.graphics.Color to HarmonyTopBar.java
- Fix float→int narrowing in HarmonyCard/GlassPanel/Sheet (explicit cast)
- Convert HarmonyOS units vp→dp, fp→sp in harmony_dimens.xml (AAPT compatibility)
- Update style-dictionary.config.mjs to auto-convert units for Android output

Build result: BUILD SUCCESSFUL, APK 18.9MB at playground/android/app/build/outputs/apk/debug/app-debug.apk

+++ platforms/android/src/main/java/com/amap/agenui/platform/harmony/widgets/HarmonyCard.java
@@ -51,7 +51,7 @@ public class HarmonyCard extends FrameLayout {
         shadowRadius = vp(16);
         shadowDx = 0;
         shadowDy = vp(2);
-        padding = vp(tokenResolver.spaceMd());
+        padding = (int) vp(tokenResolver.spaceMd());
 
         bgPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         bgPaint.setColor(tokenResolver.surfacePrimaryColor());
